+++ app/src/main/java/eu/kanade/presentation/browse/components/MigrationActionIcon.kt
@@ -17,7 +17,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
-import eu.kanade.tachiyomi.ui.browse.migration.advanced.process.MigratingManga
+import mihon.feature.migration.list.models.MigratingManga
 import tachiyomi.i18n.MR
 import tachiyomi.i18n.sy.SYMR
 import tachiyomi.presentation.core.i18n.stringResource
