+++ app/src/main/java/mihon/feature/migration/list/components/MigrationProgressDialog.kt
@@ -1,4 +1,4 @@
-package eu.kanade.presentation.browse.components
+package mihon.feature.migration.list.components
 
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.material3.AlertDialog
