+++ app/src/main/java/mihon/feature/migration/list/models/MigratingManga.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.ui.browse.migration.advanced.process
+package mihon.feature.migration.list.models
 
 import android.content.Context
 import kotlinx.coroutines.CoroutineScope
